Dyno Customer Hotline Meme
A funny interactive meme for use with Dyno.

Sadly, as Github is my kryptonite, I figure I'd just put it all into one PR and let the owner or one of the managers of this Github separate it if they like it enough (I hope that's not too much of a problem).

If you wish to use the announce command to announce the intro I used for it as well here is the intro:

Dyno Hotline is here to help (or piss you off, whichever you decide). To begin, type `?1-8002-nooblance`, then continue with `?banhammer` to find out about the `?banhammer` command, then follow the prompts.

Don't forget to use the `?creator` command to find out about the creator if you decide that you have more questions. And as always, we are here to help (or royally screw you).