Deathmatch
This command follows the format ?deathmatch [username] and challenges the specified user to a 'deathmatch' fight decided randomly. The 4 possibilities are typeracer, roulette, coin flip, and insult. As it might have repercussions for pinging people, I suggest the implementation of a role that restricts the user or receiver of the command.