Plague
There has been some controversy on this command. Please read the entire description. Requires two roles, Infected and Uninfected. (1/3 of the Infection Commands) This command gives the user the role 'Infected' if they are 'Uninfected'. The dm in the command is the bypass the {silent} and only dms the user, not others (though it will show this dm in the chat, oddly enough). The {silent} is used so that Dyno won't note the role change in the channel. Again, this command will not ping users other than the person who used it.