Infect
There has been some controversy on this command. Please read the entire description. Requires two roles, Infected and Uninfected. (1/3 of the Infection Commands) This command follows the format ?infect [username] and changes the specified user to the 'Infected' Role while removing the 'Uninfected' role. It then sends a dm to the user telling them they have been infected. I believe this is the source of the controversy, as this command allows users to ping others. This can be left as is or can be modified. For example, the command could simply not include the dm. This, however, makes the command pretty mediocre, as it does not actually tell the infected person that they have been infected (and people are unlikely to check their roles to see if they are). 

Again, the dm section of the command is to mitigate the effect of {silent}, which was added to hide the announcement Dynobot makes when roles are changed.