Cure
There has been some controversy on this command. Please read the entire description. Requires two roles, Infected and Uninfected. (3/3 of the Infection Commands). This command swaps the user from the 'Infected' role to the 'Uninfected' role. It also induces a possibility (1/2) of having the user become muted for 3 minutes, as a 'side effect'.

Note that any members of the server with the 'Infected' role do not have to use this command. There is nothing negatively affecting a person with the 'Infected' role; they are not forced to use this command. As well, I believe the 'side effect' of this command is a minor one, but it can be modified to not be included at all.

Thank you for your consideration.