Nickname Requests Commands (#41)
* Nickname Request for Users

Role Requirement: None
Usage: !nickname <your crazy cool meant to be nickname>
Description: Used to let users request a nickname for themselves, but needs to be approved.
Part one: Puts a confirmation that the request was submitted into the channel the command was used - For user
Part two: Puts a embed into a nickname log channel for a staff member to approve or deny the request - For staff

* Nickname Request Approve for Staff

Roles Required: Staff
Usage: !napp <user to nick> <cool nickname you approved>
Description: To approve the nickname request. DMs the user a message saying that the nick was approved and given. Gives nick.

* Nickname Request Deny for Staff

Role Require: Staff
Usage: !nden <user that did the request> <no so cool denied nickname>
Description: Gives DM to user that the nickname was denied.