Create report.txt
When users do `?report {UserToReport} {Evidence} {Offence}` it will send the information to #reports and mention the @ReportsTeam so they know a new report has come in. The command then goes back to #report-here and lets the user know the report was successfully created, and that the @ReportsTeam will get back to them shortly. 

This is just a basic report command, useful for moderating when staff aren't on, as users can report them easily, for staff to see when they are on.